requirements.txt
asgiref==3.11.0
cffi==2.0.0
cryptography==46.0.3
Django==6.0.1
django-extensions==4.1
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==26.0
pycparser==3.0
PyMySQL==1.1.2
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
waitress==3.0.2
whitenoise==6.7.0
